Apply Dodyg comments:
    Please modify the sample so it does not rely on a solution.
    Merge Startup.cs to Program.cs
    Remove the json configuration files and configure it via the code instead in Program.cs
    Write a readme.md on how to run the sample, e.g. npm -install, etc.
