+++ projects/dotnetify-react/Dotnetify-ReactJs/DemoVM.cs
