--- projects/dotnetify-react/Dotnetify-ReactJs/Dotnetify-ReactJs.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.106
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dotnetify-ReactJs", "Dotnetify-ReactJs\Dotnetify-ReactJs.csproj", "{30CA6613-1AC8-4253-8525-AE20973BA793}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{30CA6613-1AC8-4253-8525-AE20973BA793}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{30CA6613-1AC8-4253-8525-AE20973BA793}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{30CA6613-1AC8-4253-8525-AE20973BA793}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{30CA6613-1AC8-4253-8525-AE20973BA793}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8E6EE3E1-5954-4481-AECE-5A4FB743DBBC}
-	EndGlobalSection
-EndGlobal
