+++ projects/dotnetify-react/Dotnetify-ReactJs/package.json
