+++ projects/dotnetify-react/Dotnetify-ReactJs/readme.md
@@ -2,11 +2,11 @@
 `npm install --save-dev webpack@4.12.2`
 `npm install --save-dev webpack-cli`
 
-# Run the following command to run the asp core application.
-`dotnet run`
-
 # Run the following command to restore dependencies.
 `npm i`
+
+# Run the following command to run the asp core application.
+`dotnet run`
 
 # Use the following to command to compile and watch the typescript scripts.
 `node node_modules/webpack/bin/webpack.js --watch`
