+++ projects/dotnetify-react/Dotnetify-ReactJs/tsconfig.json
