+++ projects/dotnetify-react/Dotnetify-ReactJs/webpack.config.js
