+++ projects/dotnetify-react/Dotnetify-ReactJs/wwwroot/scripts/index/app.tsx
